+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/controller/dto/AcquirerFeeResponse.java
@@ -1,6 +1,8 @@
 package com.processing.merchantacquirer.controller.dto;
 
+import java.math.BigDecimal;
+
 public record AcquirerFeeResponse(
-        Long acquirerFee
+        BigDecimal acquirerFee
 ) {
 }
